+++ src/Lab2/Entities/Cpus/ICpu.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using Itmo.ObjectOrientedProgramming.Lab2.Entities.Cpu;
 using Itmo.ObjectOrientedProgramming.Lab2.Model.Sockets;
 
 namespace Itmo.ObjectOrientedProgramming.Lab2.Entities.Cpus;
@@ -7,8 +6,8 @@ namespace Itmo.ObjectOrientedProgramming.Lab2.Entities.Cpus;
 public interface ICpu : IBuilderDirector<ICpuBuilder>, IPcComponent,
     IPcComponentWithPowerConsumption
 {
-    public double CoreFrequency { get; }
-    public double CoreAmount { get; }
+    public int CoreFrequency { get; }
+    public int CoreAmount { get; }
     public PcSocket Socket { get; }
     public bool VideoCore { get; }
     public IReadOnlyCollection<int> RamFrequency { get; }
